requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.2
djangorestframework==3.16.0
docopt==0.6.2
dotenv==0.9.9
idna==3.10
Js2Py==0.74
packaging==24.2
pipwin==0.5.2
psycopg2==2.9.10
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
